requirements.txt
flask
flask_sqlalchemy
flask_login
flask_wtf
wtforms
requests
python-dotenv
openai
